requirements.txt
pandas
streamlit
matplotlib
tqdm
essentia-tensorflow
seaborn
numpy